% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartogony.R
\name{prjmaker}
\alias{prjmaker}
\alias{prjmaker.Extent}
\alias{prjmaker.Spatial}
\alias{prjmaker.default}
\title{Create PROJ.4 string}
\usage{
prjmaker(x, ...)

\method{prjmaker}{Extent}(x, family = "ortho")

\method{prjmaker}{Spatial}(x, family = "ortho")

\method{prjmaker}{default}(x, y, family = "ortho")
}
\arguments{
\item{x}{contextual object, or central x value for default method}

\item{...}{arguments passed to methods}

\item{family}{projection family to use}

\item{y}{central y value, for default method}
}
\value{
character PROJ.4 string
}
\description{
Create PROJ.4 string
}
\examples{
prjmaker(100, -10)
prjmaker(100, -10, family = "laea")
prjmaker()
}

