% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartogony.R
\name{projDFcolumns}
\alias{projDFcolumns}
\title{Transform data frame columns.}
\usage{
projDFcolumns(x, proj, input = c("x_", "y_"), output = c("X", "Y"),
  densify = FALSE)
}
\arguments{
\item{x}{data frame}

\item{proj}{target PROJ.4 string}

\item{input}{names of source coordinates, default \code{c("x_", "y_")}}

\item{output}{names of target coordinates, added to data frame (or modified if they exist)}
}
\description{
Source and target names (input/output) may be identical to replace the inputs.
}

